This commit includes a series of updates to the HI2UC plugin. The main change is an overhaul of the conversion logic in UppercaseConverterStrategy.cs to improve efficiency and handling of symbols and word boundaries. It also includes bug fixes in several files, including MoodCasingConverter.cs and PluginForm.Designer.cs.
Other updates encompass changing the donate button image from a .jpg to a .png for improved image quality, and some administrative changes, such as updating the copyright year and incrementing the version number from 4.2 to 4.3.

The goal of these changes is to enhance the performance and user experience of the HI2UC plugin.