Remove 'Apply' button and related methods
The 'Apply' button was removed from the plugin form, as it was deemed unnecessary. Instead of having to manually apply changes, modifications made by the user take effect immediately. This includes the resizing of the listViewFixes, previously triggered by a click on the 'Apply' button. Now, this resizing is handled by the ListViewFixes_Resize method directly. Correspondingly, the ButtonApply_Click method was also removed. Plus codestyle improvements have been made to increase the code readability and maintenance.