Refactor ListView population to be more efficient
A method within the `PluginForm` class, called `AddToListView` was excessively recreating ListView items within a loop. This task has been refactored into an extension method within the `RemoveLineBreakResult` class which converts `RemoveLineBreakResult` items directly to ListViewItems. This makes for more readable and efficient code. The extension method was housed in an appropriate class to respect the Single Responsibility Principle.