Add extension method to filter checked ListViewItems
Added an extension method "CheckItems()" to the ListView.ListViewItemCollection class. This method will return all checked items in the collection. This change simplifies and declutters the ListView item checking code in PluginForm.cs by replacing the foreach loop's inner condition with a function call to the new extension method. This makes the code cleaner and easier to read.