Refactor ListView check-uncheck methods to extension
The check, uncheck and invert check methods of the ListView in PluginForm.cs were refactored. This revision involves moving these three methods to a new ListViewExtensions class to improve code readability and reusability. Also, the language version of the project was updated to the latest to take advantage of all the modern language features of C#.