Add new enums to HIStyle
Added two new enums - 'None' and 'SentenceCase' to the 'HIStyle' enumeration in HI2UC project. The 'None' enum is introduced to help represent a null or undefined style. The 'SentenceCase' enum is added as an additional formatting option in response to user requests. Also, re-ordered the position of 'UpperLowerCase' for consistency across the file.