Add debugger launch in DoAction method
A System.Diagnostics namespace has been added and a debugger launch code block has been inserted in the DoAction method. This is to facilitate debugging during the application's startup sequence or when a debugger is not previously attached. The respective code block is conditional based on whether the application is in DEBUG mode or not.